refactor(adapters): move TelegramAdapter streaming callbacks to telegram_outbound
telegram.py was 405 lines (max 300). Extract _make_streaming_callbacks body
into build_streaming_callbacks() in telegram_outbound.py, reducing telegram.py
to 299 lines. Thin facade now delegates via a single call.
